third_party/WebKit/LayoutTests/http/tests/loading/doc-write-sync-third-party-script-conn-type-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
CONSOLE WARNING: line 18: A Parser-blocking, cross-origin script, http://localhost:8000/loading/resources/js-loaded.js?conn-type, is invoked via document.write. This may be blocked by the browser if the device has poor network connectivity.
CONSOLE WARNING: line 18: A Parser-blocking, cross-origin script, http://localhost:8000/loading/resources/js-loaded.js?conn-type, is invoked via document.write. This may be blocked by the browser if the device has poor network connectivity.
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This is a testharness.js-based test.
PASS Document.written scripts are not blocked in a fast connection 
Harness: the test ran to completion.
